third_party/WebKit/LayoutTests/media/W3C/video/networkState/networkState_during_progress-expected.txt

videoElement.networkState should be NETWORK_LOADING during progress event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "2" is "2"
PASS successfullyParsed is true

TEST COMPLETE
spec reference
